requirements.txt
pandas
numpy
openai
langchain
langchain-community
faiss-cpu
streamlit
matplotlib
python-dotenv